Inlined RecentTagVersionExtractor into only class using it
